// src/api/generated/models/adminCreateCycleResponse.ts
/**
 * 由 Orval 根据 docs/openapi.yaml 自动生成；请勿手动修改。
 */
import type { ExamCycle } from './examCycle';
import type { AdminCreateCycleResponseMessage } from './adminCreateCycleResponseMessage';

export interface AdminCreateCycleResponse {
  /**
   * 成功固定0
   * @minimum 0
   * @maximum 0
   */
  code: number;
  data: ExamCycle;
  message: AdminCreateCycleResponseMessage;
}